+++ input/input.h
@@ -18,11 +18,13 @@
 struct bsal_dna_sequence;
 struct bsal_input_operations;
 
+#include <stdint.h>
+
 struct bsal_input {
     struct bsal_input_operations *operations;
     void *implementation;
     char *file;
-    int sequences;
+    uint64_t sequences;
     int error;
 };
 
@@ -33,7 +35,7 @@ void bsal_input_destroy(struct bsal_input *input);
 int bsal_input_get_sequence(struct bsal_input *input,
                 char *sequence);
 char *bsal_input_file(struct bsal_input *input);
-int bsal_input_size(struct bsal_input *input);
+uint64_t bsal_input_size(struct bsal_input *input);
 void *bsal_input_implementation(struct bsal_input *input);
 int bsal_input_detect(struct bsal_input *input);
 int bsal_input_has_suffix(struct bsal_input *input, const char *suffix);
