+++ input/input_proxy.h
@@ -18,7 +18,7 @@ void bsal_input_proxy_init(struct bsal_input_proxy *proxy,
 int bsal_input_proxy_get_sequence(struct bsal_input_proxy *proxy,
                 char *sequence);
 void bsal_input_proxy_destroy(struct bsal_input_proxy *proxy);
-int bsal_input_proxy_size(struct bsal_input_proxy *proxy);
+uint64_t bsal_input_proxy_size(struct bsal_input_proxy *proxy);
 int bsal_input_proxy_error(struct bsal_input_proxy *proxy);
 
 void bsal_input_proxy_try(struct bsal_input_proxy *proxy,
